関数の型を変更

+++ main.c
@@ -2,7 +2,7 @@
 #include <stdlib.h> //exit関数とmalloc関数を使用するため追加
 #include <math.h> //sqrt関数を使用するため追加
 
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 受け渡しがdouble型であるためintからdoubleに変更
 {
   int y;
   if( x < 0 ){ //x <= 0からx < 0に変更
